* upstream/pr/71: (26 commits)
  CMake: for tests, accept bigger -Wframe-larger-than
  compiler-flags: forgot to pass --coverage to executable link
  RawspeedException: ThrowException(): use TLS, be thread-safe.
  Rework & de-duplicate exception code
  Add some final keywords, to help with devirtualization.
  CMake: enable LTO by default if compiler supports it
  Add some override keywords
  Add some [[noreturn]] attributes
  Cr2Decoder: create template hell
  Cr2Decoder: int sraw_coeffs[3] is private.
  RawDecoder: store threads in vector, properly init RawDecoder* parent
  DngOpcodes: init Flags mFlags by default.
  RawImage: add bogus move assignment operator.
  Point.h: iRectangle2D cleanup
  Point.h: iPoint2D cleanup
  Add some tests for iPoint2D class
  CodeCoverage: add coverage-zerocounters and coverage-html targets
  Add some tests for common.h
  DngDecoder::dropUnsuportedChunks(): rework, print warnings
  RawDecoder: refactor askForSamples(), ignore mode 'dng'.
  ...

+++ src/librawspeed/decoders/X3fDecoder.cpp
@@ -232,10 +232,10 @@ void X3fDecoder::decompressSigma( X3fImage &image )
             // Interpolate 1 missing pixel
             int blue_mid = ((int)blue[0] + (int)blue[3] + (int)blue_down[0] + (int)blue_down[3] + 2)>>2;
             int avg = dst[0];
-            dst[0] = clampBits(((int)blue[0] - blue_mid) + avg, 16);
-            dst[3] = clampBits(((int)blue[3] - blue_mid) + avg, 16);
-            dst_down[0] = clampBits(((int)blue_down[0] - blue_mid) + avg, 16);
-            dst_down[3] = clampBits(((int)blue_down[3] - blue_mid) + avg, 16);
+            dst[0] = clampBits(((int64)blue[0] - blue_mid) + avg, 16);
+            dst[3] = clampBits(((int64)blue[3] - blue_mid) + avg, 16);
+            dst_down[0] = clampBits(((int64)blue_down[0] - blue_mid) + avg, 16);
+            dst_down[3] = clampBits(((int64)blue_down[3] - blue_mid) + avg, 16);
             dst += 6;
             blue += 6;
             blue_down += 6;
